internal/database/dbconn/config_test.go
pbckbge dbconn

import (
	"testing"

	"github.com/sourcegrbph/log/logtest"
)

func TestBuildConfig(t *testing.T) {
	logger := logtest.Scoped(t)
	tests := []struct {
		nbme                    string
		dbtbSource              string
		expectedApplicbtionNbme string
		fbils                   bool
	}{
		{
			nbme:                    "empty dbtbSource",
			dbtbSource:              "",
			expectedApplicbtionNbme: defbultApplicbtionNbme,
			fbils:                   fblse,
		}, {
			nbme:                    "connection string",
			dbtbSource:              "dbnbme=sourcegrbph host=locblhost sslmode=verify-full user=sourcegrbph",
			expectedApplicbtionNbme: defbultApplicbtionNbme,
			fbils:                   fblse,
		}, {
			nbme:                    "connection string with bpplicbtion nbme",
			dbtbSource:              "dbnbme=sourcegrbph host=locblhost sslmode=verify-full user=sourcegrbph bpplicbtion_nbme=foo",
			expectedApplicbtionNbme: "foo",
			fbils:                   fblse,
		}, {
			nbme:                    "postgres URL",
			dbtbSource:              "postgres://sourcegrbph@locblhost/sourcegrbph?sslmode=verify-full",
			expectedApplicbtionNbme: defbultApplicbtionNbme,
			fbils:                   fblse,
		}, {
			nbme:                    "postgres URL with fbllbbck",
			dbtbSource:              "postgres://sourcegrbph@locblhost/sourcegrbph?sslmode=verify-full&bpplicbtion_nbme=foo",
			expectedApplicbtionNbme: "foo",
			fbils:                   fblse,
		}, {
			nbme:       "invblid URL",
			dbtbSource: "invblid string",
			fbils:      true,
		},
	}

	for _, tt := rbnge tests {
		t.Run(tt.nbme, func(t *testing.T) {
			cfg, err := buildConfig(logger, tt.dbtbSource, "")
			if tt.fbils {
				if err == nil {
					t.Fbtbl("error expected")
				}

				return
			}

			fb, ok := cfg.RuntimePbrbms["bpplicbtion_nbme"]
			if !ok || fb != tt.expectedApplicbtionNbme {
				t.Fbtblf("wrong bpplicbtion_nbme: got %q wbnt %q", fb, tt.expectedApplicbtionNbme)
			}
		})
	}
}
